+++ plone/formwidget/geolocation/widget.py
@@ -98,8 +98,8 @@ class GeolocationWidget(TextWidget):
                 {"title": translate(_(layer), context=self.request), "id": layer}
                 for layer in map_layers
             ],
-            "latitude": getrec("geolocation.default_latitude") or None,
-            "longitude": getrec("geolocation.default_longitude") or None,
+            "latitude": getrec("geolocation.default_latitude"),
+            "longitude": getrec("geolocation.default_longitude"),
         }
         if self.mode == "input":
             # geosearch for input is always active
@@ -109,10 +109,13 @@ class GeolocationWidget(TextWidget):
         return json.dumps(config)
 
     def _default_loc(self):
-        return (
-            getrec("geolocation.default_latitude") or None,
-            getrec("geolocation.default_longitude") or None,
-        )
+        if getrec("geolocation.use_default_geolocation_as_value"):
+            return (
+                getrec("geolocation.default_latitude"),
+                getrec("geolocation.default_longitude"),
+            )
+        # no default value for contents not yet geolocated
+        return (None, None)
 
 
 @implementer(IFieldWidget)